Move macos helpers to own file
